requirements.txt
django~=4.1.4
django-phonenumber-field[phonenumbers]~=7.0.1
Pillow~=9.3.0
python-dotenv~=0.21.0
python-telegram-bot==13.15
psycopg2-binary~=2.9.5
gunicorn~=20.1.0
djangorestframework-3.14.0

